Move ChartDetailCursor into own folder, put stuff related to it in there.
