+++ src/encoded/static/components/viz/ChartDetailCursor/CursorComponent.js
