+++ src/encoded/static/components/viz/ChartDetailCursor/index.js
@@ -1,0 +1,7 @@
+'use strict';
+
+var ChartDetailCursor = require('./ChartDetailCursor');
+
+export default ChartDetailCursor;
+
+export const CursorViewBounds = require('./CursorViewBounds').default;
