+++ src/encoded/static/components/viz/components/index.js
@@ -42,4 +42,3 @@ module.exports.ZoomCursor = require('./ZoomCursor');
  */
 module.exports.RotatedLabel = require('./RotatedLabel');
 
-module.exports.CursorComponent = require('./CursorComponent');
